+++ src/Http/Controllers/HealthController.php
@@ -3,13 +3,14 @@
 namespace Butler\Service\Http\Controllers;
 
 use Butler\Health\Repository;
+use Illuminate\Http\Middleware\SetCacheHeaders;
 use Illuminate\Routing\Controller as BaseController;
 
 class HealthController extends BaseController
 {
     public function __construct()
     {
-        $this->middleware('cache.headers:no_store,no_cache,must_revalidate');
+        $this->middleware(SetCacheHeaders::using('no_store'));
     }
 
     public function __invoke(Repository $repository)